chore(test): changed imports in tests
