+++ commit_helper/utils/file_handler.py
@@ -3,10 +3,10 @@ from yaml import YAMLError
 from os import system
 # utils imports
 from .utils import create_file
-from .utils import debug
-from .utils import get_text
-from .utils import get_context
 from .utils import gen_co_author
+from .text_utils import debug
+from .text_utils import get_text
+from .text_utils import get_context
 # conventions imports
 from commit_helper.conventions.karma_angular import angular_convention
 from commit_helper.conventions.changelog import changelog_convention
@@ -14,7 +14,7 @@ from commit_helper.conventions.symphony_cmf import symphony_convention
 from commit_helper.conventions.no_convention import just_message
 
 
-def handle_file_based_commit(file_path, debug_mode):
+def handle_file_based_commit(file_path, debug_mode, args):
     with open(str(file_path), 'r') as stream:
         try:
             config = safe_load(stream)
