+++ commit_helper/utils/flag_commit_handler.py
@@ -1,10 +1,10 @@
 from os import system
 # utils imports
 from .utils import create_file
-from .utils import debug
-from .utils import get_text
-from .utils import get_context
 from .utils import gen_co_author
+from .text_utils import debug
+from .text_utils import get_text
+from .text_utils import get_context
 # conventions imports
 from commit_helper.conventions.karma_angular import angular_convention
 from commit_helper.conventions.changelog import changelog_convention
